Don't allow sub-second travel times.
Doesn't happen on my machine, but it seems that sometimes we end up
with zeroish startup times rounding down to negative start times.
Since it doesn't happen on my machine I'm not sure, but this is a
guess at a fix.

I've also removed all of the rounding from the UI so that it might be
more obvious what's wrong when it happens.