Fixes bug when you blow up the Kuz
So I blew up the Kuznetsov in my run through and apparently that causes
a game state corruption somewhere in the ground planning phase. Re-generating the ground planners seems to override this bug. I'm sure there's a better way to fix this at the root cause, but I'm not 100% sure where that is because I've already blown up the Kuz.